Bump versioncode
The versionName is actually just a user facing string with no implied meaning. What determines if an apk is an update, downgrade or the same version is actually the versionCode: an integer that gets raised with each release (some even raise it for pre-releases). For further information, please refer to https://developer.android.com/studio/publish/versioning.html.